linux: turn on -Wextra
This is a followup to an earlier change (r38266) which did most of the
warning-related cleanup.  This one just flips the flag, and fixes some
new warnings that crept in during the window while the flag was off.

Second try, now with some libpng fixes.

BUG=34160

Review URL: http://codereview.chromium.org/1320011

git-svn-id: http://src.chromium.org/svn/trunk/src@42700 4ff67af0-8c30-449e-8e8b-ad334ec8d88c