Work around coordinate problems in page zoom for Mac plugins
This is a temporary fix until the underlying problem with coordinates in windowless plugins is fixed at the WebKit level, since it's a more serious problem on the Mac (due to all plugins being windowless). Using what we know about the plugin location on screen, we infer zoom levels from the coordinates and correct the coordinates (to within a pixel or two, which is the best we can do with the already-rounded values we are given).

BUG=26466
TEST=Zoom a page with plugins on the Mac; mouse interaction should work correctly.

Review URL: http://codereview.chromium.org/1594001

git-svn-id: http://src.chromium.org/svn/trunk/src@43106 4ff67af0-8c30-449e-8e8b-ad334ec8d88c