Removing a clock dependent behavior in testserver which caused TestTwoAuths and TestDigestAuth to fail when they take more than 10 seconds to send auth (which happens sometimes under valgrind).
Now the unittest can control whether it wants to receive a stale
nonce reply by requesting /auth-digest/stale. I plan to add a test
for the stale flow in a subsequent patch.

Also, this patch cleans up NavigateToURL calls in LoginPromptTest
so that they are all in-place ASSERTs. This makes the messages more
useful.

BUG=36163,25794,38580
TEST=sh tools/valgrind/chrome_tests.sh -t ui --gtest_filter=LoginPromptTest.*
Review URL: http://codereview.chromium.org/1508001

git-svn-id: http://src.chromium.org/svn/trunk/src@43266 4ff67af0-8c30-449e-8e8b-ad334ec8d88c