Add batchAdd functions to JS List
BMM can be really slow to render when there are hundreds of bookmarks in the
bookmark bar.  This will hide the list while it's rendering the new elements
which increases performance by a lot since it doesn't have to do any
measurements while it's adding all the elements.

BUG=39865
TEST=Load BMM with over 500 bookmarks in the bookmark bar

Review URL: http://codereview.chromium.org/1540014

git-svn-id: http://src.chromium.org/svn/trunk/src@43543 4ff67af0-8c30-449e-8e8b-ad334ec8d88c