Win32 plumbing for dragend event.
This plumbs the actual dropEffect that occurred back into WebKit so that dragend is dispatched with the correct dropEffect. Note that this only helps with drags that start in Chrome and end outside of Chrome. Since WebDropTarget always lies and claims that no drop occurred, drags that start in Chrome and end in Chrome will still be incorrect.

BUG=39399
TEST=Manual testing using the attached test case on the bug.

Review URL: http://codereview.chromium.org/1136007

git-svn-id: http://src.chromium.org/svn/trunk/src@43726 4ff67af0-8c30-449e-8e8b-ad334ec8d88c