Don't clear VideoRendererImpl's proxy_ variable during OnStop().
Nasty threading issue involving PipelineThread, VideoRendererThread and RenderThread:
  - VideoRendererThread continuously calls OnFrameAvailable(), as expected
  - RenderThread gets the signal to destroy a <video>
  - RenderThread blocks on PipelineThread to finish clean up
  - PipelineThread calls OnStop(), which sets proxy_ to NULL
  - Race condition occurs if VideoRendererThread was already on its way to call OnFrameAvailable() again

Now we could fix this by adding more locks (i.e., inside VideoRendererBase::ThreadMain() before we call OnFrameAvailable()), but that defeats the entire point of keeping VideoRendererThread as lock-free as possible.

We could also add an extra lock in VideoRendererImpl to synchronize access to proxy_, but that's unnecessary as proxy_ outlives VideoRendererImpl so there's no need to set it to NULL.  Furthermore it's OK to keep calling proxy_->Repaint() since that will post tasks to the RenderThread, which is conveniently blocked on PipelineThread to finish, who is blocked on VideoRendererThread to finish calling proxy_->Repaint()!  After that task is posted, everything unblocks and terminates and when RenderThread resumes it notifies proxy_ to cancel any pending tasks (such as that last repaint).

BUG=35858
TEST=layout tests should crash less frequently on shutdown

Review URL: http://codereview.chromium.org/1625003

git-svn-id: http://src.chromium.org/svn/trunk/src@43766 4ff67af0-8c30-449e-8e8b-ad334ec8d88c