Further improvments in IE switching
Manually loading active document in IMoniker::BindToObject does
not work in IE6. The binding is not initialized and directly calling
BindToStorage on it just does not return the data.
The solution is to register ActiveDoc's CLSID as handler for 'text/html'
mime type in the bind context marked for switching. Urlmon will correctly
instantiate it them

BUG=none
TEST=covered by existing tests.

Review URL: http://codereview.chromium.org/1595021

git-svn-id: http://src.chromium.org/svn/trunk/src@44274 4ff67af0-8c30-449e-8e8b-ad334ec8d88c