Fix for remaining major issues with moniker patch
This patch fixes things so that they now work 'naturally', i.e switch
to chrome frame looks just like switching to pdf. The same bind context
and underlying network transaction objects are used and requests aren't
sent multiple times. Finally the TestPostReissue test is happy!

This patch fixes two issues with earlier scheme:
Issue 1: data can be read only once from a given bind context. Mshtml
looks at the mime type in the FORMATETC but never issues a read. So when
it switches, the new active document can read from start. Since we are
reading the data for switching, when we actually switch, we have to 
transfer over the data read so far to our UrlmonUrlRequest.

Issue 2: The very first UrlmomUrlRequest has to start binding right away,
in the context of BindToStorage or network transcations associated with
the bind context are gone (due to OnStopBinding for BindToObject).

TEST=covered by existing tests
BUG=33332

Review URL: http://codereview.chromium.org/1508033

git-svn-id: http://src.chromium.org/svn/trunk/src@44427 4ff67af0-8c30-449e-8e8b-ad334ec8d88c