Start up extension event routers later in BrowserMain, so bookmarks is loaded after bookmark import in first run.
I am not familiar with the extension code. It seems currently
ExtensionHistoryEventRouter and ExtensionAccessibilityEventRouter
are always initialized while ExtensionBrowserEventRouter
and ExtensionBookmarkEventRouter are initialized on demand. Since
we have a build in bookmark manager extension, I start all event
routers by default in this change. Let me know if this is an
issue.

BUG=40144
TEST=chrome.exe --first-run

Review URL: http://codereview.chromium.org/1578032

git-svn-id: http://src.chromium.org/svn/trunk/src@44668 4ff67af0-8c30-449e-8e8b-ad334ec8d88c