posix: handle the return value of close() in more places.
Generally, we don't expect it to fail and there isn't much we
can do anyway, but it's good to at least consider each case and
do something so we can continue to receive warnings in situations
where we forgot to check the return code.

(Bonus extra bugfix: use int where we previously had bool.
I think it compiles to the same thing but the old code was
definitely wrong.)

Review URL: http://codereview.chromium.org/1564037

git-svn-id: http://src.chromium.org/svn/trunk/src@44844 4ff67af0-8c30-449e-8e8b-ad334ec8d88c