Defer WidgetGtk's focus manager destruction
This is for the case when a WidgetGtk is hosted in another WidgtGtk and
holds a reference to its focus manager. When hosting widget is destroied,
it gets the "destroy" signal before the contained widget and
thus schedules its destruction before the contained widget. And currently,
we are not clearing the focus manager reference in another root view and
when contained widget destructs, we crash. By changing focus manager's
release into another message, the focus manager remains valid during
contained widget destruction.

BUG=none
TEST=In ChromeOS, bring up options dialog and then click on buttons such as language settings, networ config etc. Then dismiss the options dialog and Chrome should not crash.

Review URL: http://codereview.chromium.org/1661003

git-svn-id: http://src.chromium.org/svn/trunk/src@45071 4ff67af0-8c30-449e-8e8b-ad334ec8d88c