In ChromeFrame with the IMoniker patch on, we should save away the redirect information received for the initial pending request. This is to ensure that we inform Chrome about this redirect which can then be followed correctly from Chrome.
The other change is to save away the URL being navigated to in the bind context context object. This is also updated if we receive a redirect
notification in our callback wrapper. The active document refers this URL if it is available and falls back to the old mechanism of retrieving
the URL from the moniker if not.

This ensures that optin urls continue to work correctly if the initial page is redirected.

Review URL: http://codereview.chromium.org/1523028

git-svn-id: http://src.chromium.org/svn/trunk/src@45114 4ff67af0-8c30-449e-8e8b-ad334ec8d88c