Close a newly opened (by cmd-clicking on a link) tab if it resulted in a download.
To test this, the CL adds an automation message AutomationMsg_NavigateAsyncWithDisposition, and a method NavigateToURLAsyncWithDisposition to TabProxy.

The only functional change is in TabContents::OnStartDownload, the rest of the changes is for testing.

BUG=10764
TEST=DownloadTest.CloseNewTab
Manual test: go to http://build.chromium.org/buildbot/continuous/mac/LATEST/, open the link "chrome-mac.zip" in a new tab. The tab should close as soon as the download starts.

Review URL: http://codereview.chromium.org/1151007

git-svn-id: http://src.chromium.org/svn/trunk/src@45158 4ff67af0-8c30-449e-8e8b-ad334ec8d88c