Adding ExtensionPrefs methods for storing update-when-idle data.
This includes adding a few methods to ExtensionPrefs that I'm going to need to
use inside ExtensionUpdater for implementing the "do updates at idle" feature.
Instead of adding more call-through stubs to the interface that 
ExtensionsService providers to the ExtensionUpdater, I instead decided it was 
time to do some refactoring to expose a ExtensionPrefs getter in the interface.

TEST=(Should be covered by unit/browser tests)
BUG=37971


Review URL: http://codereview.chromium.org/1695018

git-svn-id: http://src.chromium.org/svn/trunk/src@45738 4ff67af0-8c30-449e-8e8b-ad334ec8d88c