This is the second side of a multi-sided webkit patch that will allow experimental window.open feature strings to be captured and passed to the client.
It should wait for: https://bugs.webkit.org/show_bug.cgi?id=38013 to land. This is required because this patch needs to use the WebWindowFeature binding struct in WebKit.

Once this is landed, a webkit patch to change the call signature for WebViewClient::createView to include the WebWindowFeatures will be submitted (and after that another chromium patch to remove the old call signature).

This patch implements old and new WebViewClient::createView signatures, and additionally passes the vector all the way to RenderViewHostDelegateHelper who will eventually use it to observe the 'background' feature.

BUG=41275
TEST=all tests should pass

Review URL: http://codereview.chromium.org/1758004

git-svn-id: http://src.chromium.org/svn/trunk/src@45877 4ff67af0-8c30-449e-8e8b-ad334ec8d88c