Reland r46049: [Mac] Make exceptions dialog a sheet.
Now with added unit tests!

XIB changes (applying to ContentExceptionsWindow.xib and GeolocationExceptionsWindow.xib):
The window is now drawn in Aqua style, with window margins, Aqua buttons and text labels for them.

At the bottom right of the window there is an additional button for closing the sheet, hooked up to the |doneButton_| outlet in the file's owner and with action |closeSheet:|.

The lefthand side buttons are inside a |GTMWidthBasedTweaker|, as is the done button. The actual tweaking is done by a newly added |GTMUILocalizerAndLayoutTweaker|.

BUG=38021
TEST=Preferences > Under the Hood > Content Settings > Exceptions should be a sheet.

Review URL: http://codereview.chromium.org/1817003

git-svn-id: http://src.chromium.org/svn/trunk/src@46225 4ff67af0-8c30-449e-8e8b-ad334ec8d88c