Put version and boot times on login screen.
BUG=chromium-os:2793, chromium-os:2980
TEST=None

When we went to login images we lost the version on the main screen.
This change moves it to the background so all wizard panels will show it.
Also we currently show the time it took to boot the system. We may change
this later but it's useful for now to keep us aware of regressions.
In addition, any issues with loading libcros are displayed here as well.

Review URL: http://codereview.chromium.org/1827001

git-svn-id: http://src.chromium.org/svn/trunk/src@46358 4ff67af0-8c30-449e-8e8b-ad334ec8d88c