Lands http://codereview.chromium.org/1744011/show for Hans:
Make reopened tabs with "about:" URLs display properly. 

- Rename TabNavigation::url_ to virtual_url_. 
Make it more explicit that a TabNavigation stores the *virtual* URL of
a 
NavigationEntry. 

- Translate virtual URLs when restoring tabs. 
Instead of putting the virtual URL in the URL proper field when
creating a 
NavigationEntry from a TabNavigation, put it in the virtual URL field,
and 
then translate this to a proper URL in 
NavigationController::CreateNavigationEntriesFromTabNavigations(). 
This fixes bug 31905. 

Contributed by Hans Wennborg <hans@chromium.org> 

BUG=31905
TEST=ui_tests --gtest_filter=TabRestoreUITest.RestoreTabWithSpecialURL*

Review URL: http://codereview.chromium.org/1946002

git-svn-id: http://src.chromium.org/svn/trunk/src@46406 4ff67af0-8c30-449e-8e8b-ad334ec8d88c