Committing on behalf of dtseng.
We currently only set focus to the widget's hWnd, but we really want
to also notify assistive technologies that a child view received focus
so that it can know to obtain an IAccessible for that custom child
view using the widget hWnd. The widget hWnd has our implementation of
IAccessible through AccessibilityWrapper/ViewAccessibility.

Original code review: http://codereview.chromium.org/1985002

BUG=43363
TEST=none


git-svn-id: http://src.chromium.org/svn/trunk/src@46687 4ff67af0-8c30-449e-8e8b-ad334ec8d88c