sync: only add to the dirty_metahandles index when marking an item dirty, rather than for every item in a WriteTransaction originals array, and make sure to restore items into dirty_metahandles on save changes failure to avoid data loss.
This prevents a situation where after a SaveChanges, a "clean" (not dirty) item will pass the
SafeToPurgeFromMemory filter because is_dirty is false, but it is in fact still in dirty_metahandles,
and we yank it from metahandles_index.  This causes a crash later on as GetByHandle will return NULL
for items in this boat.  The other part of this fix makes sure we don't lose a bookmark change on
browser restart due to a single sqlite operation failing during SaveChanges by restoring items into the 
dirty index.

BUG=43047,40341

Review URL: http://codereview.chromium.org/2043002

git-svn-id: http://src.chromium.org/svn/trunk/src@46712 4ff67af0-8c30-449e-8e8b-ad334ec8d88c