ChromeFrame would not switch to Chrome at times. The navigation manager in ChromeFrame maintains the current top level URL which is used by the moniker patch to decide whether to wrap the callback. This url is typically set in the Bho in the BeforeNavigate callback. We also had code to update this url when we receive a navigation update from Chrome. The reason we would not switch is because of a race condition between receiving a top level navigation request from chrome and receiving a navigation update for the previous operation which caused us to not wrap the bind status callback.
Fix is to not update the url in the callback when we receive a navigation status update. In any case it is Chrome's responsibility to inform
ChromeFrame about top level url navigations. We should fix the bug if any in Chrome which caused this hack to go in.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=43922

Other changes are to replace the NOTREACHED in the chrome frame automation proxy code with a DLOG as this fires during chrome frame test runs.

Bug=43922

Review URL: http://codereview.chromium.org/1991008

git-svn-id: http://src.chromium.org/svn/trunk/src@47086 4ff67af0-8c30-449e-8e8b-ad334ec8d88c