Implement IndexedDatabase::open up until re-entrance into WebKit.  That (and the rest of IndexedDB) will be done in subsequent patches.  Included in this patch is a lot of other infrastructure for making IndexedDB work.  Included is a conversion from the DOMStorageDispatcherHost into a dispatcher host for all APIs that have a backend in WebKit.
I named it WebKitAPIDispatcherHost.  Since it's in browser/in_process_webkit and
it is for APIs that connect up to the WebKit API and it's used for APIs whose
backend is implemented in WebKit I thought the name was decent and not _too_
confusing, but if you have better ideas, please let me know.

This includes some code that you've already reviewed (darin), but a lot has
changed so please take a look at all of it.

TEST=Not much to test yet + behind a flag.
BUG=none

Review URL: http://codereview.chromium.org/1599009

git-svn-id: http://src.chromium.org/svn/trunk/src@47268 4ff67af0-8c30-449e-8e8b-ad334ec8d88c