AutoFill filling billing address and company name from profile.
The lego.com page represents a couple different issues.  (1) is that the address fields are identified as billing address fields.  So when filling from a profile and not a credit card those fields did not get filled.  Based on discussion with Jeff Chang we're changing this so that when filling a profile, billing fields get treated as normal address fields and will be filled.  In the code the distinction of billing versus "home" address information has now been removed as it is no longer applicable.  (2)  Company fields now get heuristically matched and filled.

BUG=40683
TEST=Manual testing of https://shop.lego.com/Checkout/Address/, \
Unit tests: AddressFieldTest.ParseCompany, AddressFieldTest.ParseCompanyEcml \
FormStructureTest.HeuristicsWithBilling FormStructureTest.HeuristicsSample6

Review URL: http://codereview.chromium.org/2085006

git-svn-id: http://src.chromium.org/svn/trunk/src@47331 4ff67af0-8c30-449e-8e8b-ad334ec8d88c