[Mac] Implement NSObject zombies.
Apple's NSZombieEnabled setting makes it easier to catch messages to
freed objects, but is mostly only useful in debugging environments.
This implements a facility like NSZombieEnabled with the following
additions:

 - The number of outstanding zombies can be configured.
 - Classes can opt-in to becoming zombies.
 - C++ destructors are correctly called on Leopard.

The goal is to allow us to enable zombies in certain production builds
to help debug some of the message-after-free bugs we have.

BUG=35590,24987
TEST=everything

Review URL: http://codereview.chromium.org/660411

git-svn-id: http://src.chromium.org/svn/trunk/src@47674 4ff67af0-8c30-449e-8e8b-ad334ec8d88c