GTK: Use label colors for our tinted buttons.
Decisions to tint the buttons the color of the selection color was made back
in the Hardy era, in a desire to match the Windows theming system, and this
worked OK with the themes as they were written at the time. In Lucid, this
means too light or too dark buttons. Instead of trying to match the
selection color, try to match the label color. This works well in all the
themes I've tried so far on Lucid.

This kind of sucks that it's not as colorful, but the buttons are now as
legible as text.

BUG=43853
TEST=none

Review URL: http://codereview.chromium.org/2159003

git-svn-id: http://src.chromium.org/svn/trunk/src@48107 4ff67af0-8c30-449e-8e8b-ad334ec8d88c