AutoFill <dl> label scraping and dynamic html in forms.
The mcphee.com site demonstrates two issues.  One is label scraping for forms that use <dl> <dt> and <dd> elements to partition labels and input elements.  The other is dynamic html where JavaScript is used to manipulate the form elements dynamically as the user proceeds through checkout.  These changes address both.  Label scraping is implemented for <dl> <dt> and <dd> elements.  Form filling, both at the FormManager and AutoFillManager levels are now more robust in the face of changing form structure between the time of page load and form fill.

This does not "fix" the mcphee page though, as we still do not fill the elements that are added to the page after the initial page load.

BUG=44323
TEST=AutoFillManagerTest.FormChangesRemoveField, AutoFillManagerTest.FormChangesAddField, FormManagerTest.LabelsInferredFromTableCellNested, FormManagerTest.LabelsInferredFromDefinitionList, FormManagerTest.FillFormChangedFormDataFields, manual tests using mphee.html and mcphee2.html sample files attached to bug report and original bug steps using live mcphee checkout.

Review URL: http://codereview.chromium.org/2240004

git-svn-id: http://src.chromium.org/svn/trunk/src@48429 4ff67af0-8c30-449e-8e8b-ad334ec8d88c