GTK: Fix padding of the page/app menus.
In r47478, I removed the hbox that the page and app menus were in because I
though that with the new left/right toolbar boxes, they were unnecessary.
But they're needed to group the two buttons together with no space between
them, but with padding on the right.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/2375002

git-svn-id: http://src.chromium.org/svn/trunk/src@48522 4ff67af0-8c30-449e-8e8b-ad334ec8d88c