posix: set thread names
There's a non-portable function for doing this: pthread_setname_np.
It's supported by OS X >= 10.6 and the Xcode debugger will show the thread
names if they're provided.  On Linux, support has just recently been added
to glibc for the same function.  Since OS coverage of the function is so
spotty, we look for the symbol at runtime; on Linux we fall back to another
implementation of the same functionality if the function isn't available.

Review URL: http://codereview.chromium.org/2774001

git-svn-id: http://src.chromium.org/svn/trunk/src@49212 4ff67af0-8c30-449e-8e8b-ad334ec8d88c