Take Keystone keys out of the framework's Info.plist. They now live only in the outer application's Info.plist.
This ensures that the framework and other contents of the versioned directory
can be bit-for-bit identical for a given version, regardless of configured
Keystone channel. This will enable binary diff updates, because they will not
need knowledge of the channel a user is upgrading from.

BUG=45017
TEST=Keystone still functions properly, channel still shows up in the about
     window and about:version of official builds
Review URL: http://codereview.chromium.org/2791001

git-svn-id: http://src.chromium.org/svn/trunk/src@49334 4ff67af0-8c30-449e-8e8b-ad334ec8d88c