Silently discard invalid OnFollowDeferredRedirect requests sent by the renderer to the browser.
It is a bug if the renderer sends these, but this check prevents the browser from crashing as a consequence of it.
One situation where this can happen today (and needs to be fixed renderer-side), is when the renderer cancel's a request in response to IGNORE policy, but then later tries to follow the redirect on the cancelled request.

BUG=20470


Review URL: http://codereview.chromium.org/2793001

git-svn-id: http://src.chromium.org/svn/trunk/src@49421 4ff67af0-8c30-449e-8e8b-ad334ec8d88c