Make timer-driven CA plugins less wasteful, and up the timer frequency
Don't do cross-process IOSurface work if nothing has changed in the layer. Now that it's less expensive, up the frequency of the timer a bit to minimize the chance of dropped frames due to drift if the plugin is updating at a different speed.

BUG=None
TEST=Core Animation plugins on the Mac still work.

Review URL: http://codereview.chromium.org/2847003

git-svn-id: http://src.chromium.org/svn/trunk/src@49800 4ff67af0-8c30-449e-8e8b-ad334ec8d88c