Added BackgroundContentsService to manage lifecycle of BackgroundContents.
If --restore-background-contents flag is passed, stores the URLs of running
BackgroundContents in preferences so they can be re-launched when the browser
restarts.

Moved logic to shutdown background contents into BackgroundContentsService so
we can use this to coordinate when to keep the browser process running.

BUG=43382
TEST=new tests

Review URL: http://codereview.chromium.org/2104018

git-svn-id: http://src.chromium.org/svn/trunk/src@50329 4ff67af0-8c30-449e-8e8b-ad334ec8d88c