Push change for stevenjb:
Issue 2848016: Store wifi network in network_menu_button.cc as a path instead of as a struct.
Store wifi network in network_menu_button.cc as a path instead of as a struct. 
This should address a potential edge case where wifi network information gets updated after the menu is opened and before a network is selected. 

BUG=None 
TEST=None
Review URL: http://codereview.chromium.org/2817036

git-svn-id: http://src.chromium.org/svn/trunk/src@51064 4ff67af0-8c30-449e-8e8b-ad334ec8d88c