Don't set the URL on folders, even to an empty string.
This suppresses a harmless LOG(ERROR) on SPECIFICS / SERVER_SPECIFICS not
matching.  That's because the server doesn't return preserve a URL field on
folder entities, so if we set it locally (even to the empty string), it won't be
returned in the result message.

BUG=45610
TEST=create a folder, rename it, verify no 'LOG(WARNING) << "Specifics mismatch"' from syncer_util.cc

Review URL: http://codereview.chromium.org/2904013

git-svn-id: http://src.chromium.org/svn/trunk/src@52269 4ff67af0-8c30-449e-8e8b-ad334ec8d88c