Revert 52441 - Revert Win specific elements of 46492 to look for perf impact
There was a regression in the moz page cycler around when
landed. This was ONLY a single core regression, which
suggests it was time spent in the browser becoming 
critical path. This change involved font layout in the
browser, and so it MIGHT be related.

I'll land this, let the per bots start their run, and 
then revert.


TBR=pkasting
Review URL: http://codereview.chromium.org/2895017

TBR=jar@chromium.org
Review URL: http://codereview.chromium.org/3018001

git-svn-id: http://src.chromium.org/svn/trunk/src@52456 4ff67af0-8c30-449e-8e8b-ad334ec8d88c