Implements policy to enable or disable the password manager, as found in the Options dialog, in the Under the Personal Stuff section.
xib changes: added an enabled binding to the policy-controlled checkboxes in the "Personal Stuff" preference pane so that they can be disabled when their value is overridden by policy. The methods for the bindings are in the preference window controller.

BUG=44395
TEST=ConfigurationPolicyPrefStoreTest.* and ConfigurationPolicyProviderWinTest.*  Manual tests: Set the policy externally.  Check that the radio buttons and the "Show saved passwords" button in Options/Personal Stuff/Passwords are disabled.  Also check that the check box is set/unset as done by the policy.

Review URL: http://codereview.chromium.org/2848039
Patch from Jean-Luc Brouillet <jeanluc@google.com>.

git-svn-id: http://src.chromium.org/svn/trunk/src@52637 4ff67af0-8c30-449e-8e8b-ad334ec8d88c