This checkin contains 2 new types of sync integration tests.
The first type of tests ("multiple client") creates a small number of clients (3) that make simultaneous, sometimes conflicting updates to a sync item and verifies that the system reaches a sane state in spite of racy conditions.

The second type of tests ("many client") creates several clients (10) that all sync to the same account and process updates sent out due to one or more of them syncing a change, and verifies that the system reaches a known sane state at the end of the test.

BUG=49035
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/2897017

git-svn-id: http://src.chromium.org/svn/trunk/src@52724 4ff67af0-8c30-449e-8e8b-ad334ec8d88c