[Mac] Adjust omnibox decoration dragging to keep image under mouse.
LocationBarDecoration::GetDragImageFrame() added to let decorations
specify where their GetDragImage() image is in the decoration.  The
mouse-down code ensures that the mouse drags from the expected point
within that area, or centeres the image under the mouse if the drag is
from somewhere else in the decoration.

Additionally tracked down why -isFlipped affected the drag point and
rewrote the code to be -isFlipped agnostic.

BUG=49102, 40771
TEST=Goto https://www.thawte.com/.  Drag from ev text, lock icon should be under cursor.
TEST=drag from location or lock icon should drag from point in icon mouse click was in.

Review URL: http://codereview.chromium.org/2870059

git-svn-id: http://src.chromium.org/svn/trunk/src@52959 4ff67af0-8c30-449e-8e8b-ad334ec8d88c