Download code cleanup:
- choose better names for some helper methods
- move code to less random places

This change also adds bigger tuples support, up to Tuple8 in base/tuple.h.
The plan is to stop using such big number of parameters, but for now it's not
trivial.

This change also fixes some UI tests, not sure why it is so.

TEST=unit_tests, browser_tests, ui_tests
BUG=48913

Review URL: http://codereview.chromium.org/2805091

git-svn-id: http://src.chromium.org/svn/trunk/src@53053 4ff67af0-8c30-449e-8e8b-ad334ec8d88c