Make ResourceDispatcherHostRequestInfo own SSLClientAuthHandler and LoginHandler references
This avoids problems where the objects disappear while the
ResourceDispatcherHostRequestInfo still holds pointers. In particular,
SSLClientAuthHandler handler adds and removes a reference to itself when the
user dismisses the dialog. ResourceDispatcherHost, however, still holds a
pointer and may attempt to notify it of a cancelled request. This stray write
may NULL out a useful field and crash.

BUG=49352
TEST=steps in bug report

Review URL: http://codereview.chromium.org/3043008

git-svn-id: http://src.chromium.org/svn/trunk/src@53082 4ff67af0-8c30-449e-8e8b-ad334ec8d88c