Revert 53849 - Implement most of the startup page controls in DOMUI prefs
Hooks up display of startup pages, removal of pages, and setting the startup pages to the currently open tabs.
Uses a select for now, but will eventually be replaced with a List so the favicons can be displayed.

BUG=48713
TEST=Startup pages can be viewed, removed, and set to current in DOMUI prefs.

Review URL: http://codereview.chromium.org/3046025

TBR=stuartmorgan@chromium.org
Review URL: http://codereview.chromium.org/3058018

git-svn-id: http://src.chromium.org/svn/trunk/src@53857 4ff67af0-8c30-449e-8e8b-ad334ec8d88c