Speculative fix for crash in DOMUIThumbnailSource.
DOMUIThumbnailSource objects are deleted on the IO thread, while they are used
on the UI thread. The DataSource documentation says that they should not live
on the IO thread, but for almost all DataSources, the only reference held is
the one by ChromeURLDataManager, which lives on the IO thread. Since I had a racy
stack where the object was being used on the UI thread while its destructor was
running on the IO thread, forcing destruction on the UI thread should fix
the crash.

Pretty much everything but changing the templated base class of DataSource to
always DeleteOnUIThread is my usual cleanup.

BUG=34115
TEST=none

Review URL: http://codereview.chromium.org/3061009

git-svn-id: http://src.chromium.org/svn/trunk/src@53873 4ff67af0-8c30-449e-8e8b-ad334ec8d88c