Disable run-time flag for device orientation in test_shell
Device orientation runtime flag must be switched off in test_shell,
as it doesn't provide a DeviceOrientationClient to the Page constructor.
WebKit r64270 triggered the problem by using this NULL client
in the DeviceOrientationController constructor, causing it to crash.
The DeviceOrientationController does not get instantiated when the runtime flag is off.

BUG=None
TEST=test_shell should not crash

Review URL: http://codereview.chromium.org/3066015

git-svn-id: http://src.chromium.org/svn/trunk/src@54125 4ff67af0-8c30-449e-8e8b-ad334ec8d88c