Make prefetching default to disabled
My earlier CL jumped the gun a bit; prefetching isn't ready to be the default,
so this CL reverses that.  Now you have to explicitly turn prefetching on to have it.

Note that prefetching was never enabled in trunk: because I haven't yet flipped the
build default for webkit to include prefetching, the command line flag had no effect
on trunk chrome.

BUG=none
TEST=chrome --enable-prefetch http://gemal.dk/browserspy/prefetch.php

Review URL: http://codereview.chromium.org/2847093

git-svn-id: http://src.chromium.org/svn/trunk/src@54485 4ff67af0-8c30-449e-8e8b-ad334ec8d88c