Don't leak memory when displaying a SFChooseIdentityPanel
SFChooseIdentityPanel has rather non-standard (buggy) referencing semantics as
compared to other sheets. In particular, it fails to take a reference to itself
while open. It also takes a reference to its delegate, introducing a
reference cycle in our implementation.

This removes the panel reference in the delegate, in light of the second
behavior, and works around problems caused by the first with an autorelease.

Also avoids an indexing error in release mode should a DCHECK fail.

R=wtc,mark
BUG=none
TEST=displaying certificate selections still work in OS X

Review URL: http://codereview.chromium.org/3044036

git-svn-id: http://src.chromium.org/svn/trunk/src@54588 4ff67af0-8c30-449e-8e8b-ad334ec8d88c