[ChromeOS] When doing a stub-login for chromeos, avoid the already-logged-in-user path
Before r54950, we used the kLoginUser switch to indicate that chrome
had crashed, and we were relaunching it for the indicated user.
http://src.chromium.org/viewvc/chrome?view=rev&revision=54950 introduced a
new semantic: if passed along with kLoginPassword, this flag now indicated
that a mocked-out login should be attempted with the given credentials.

Doing this as well as starting up like we were relaunching the browser for
the indicated user caused some DCHECKs to fire.

BUG=None
TEST=do a stub login, watch it succeed

Review URL: http://codereview.chromium.org/3026053

git-svn-id: http://src.chromium.org/svn/trunk/src@55127 4ff67af0-8c30-449e-8e8b-ad334ec8d88c