Reverse rows in DecoderZlib for remoting
Windows images are upside down, so when we decode the stream and paint the
image we need to reverse the rows.

This is a temporary solution since it sets to reverse rows by default. We'll
move to use the information from the update stream to determine we should
reverse or not.

Review URL: http://codereview.chromium.org/2883055

git-svn-id: http://src.chromium.org/svn/trunk/src@55314 4ff67af0-8c30-449e-8e8b-ad334ec8d88c