Factor out command-line quoting code on Windows.
Our command line code is not very good with respect to getting
quoting right.  Because of this, callers will sometimes themselves
attempt to quote arguments, though some of them get it wrong(!) by
just surrounding with quotes.

The first step is to add a quoting function that we think is correct
and update the unit test to test it.

(Note that most of this is Windows-specific, because on POSIX
we can pass arguments to commands as a vector -- trying to do
quoting on POSIX is usually wrong.)

Review URL: http://codereview.chromium.org/3007038

git-svn-id: http://src.chromium.org/svn/trunk/src@55591 4ff67af0-8c30-449e-8e8b-ad334ec8d88c