sync: Fix backend configuration to support sync db refresh.
Removes the has_new bit in favor of checking if initial_sync_ended for requested types, and recreates the sync db on corruption (as we used to).  These together mean we refresh the sync db on corruption.  Also added some defensive dchecks.

To test this, I ripped out the test-only #ifdefs in SyncBackendHost so the tests use the real Initialize code path, and added a toolbox of stuff to control when we set initial sync ended bits during testing.

BUG=50965
TEST=DirectoryBackingStoreTest, ProfileSyncService* tests.

Review URL: http://codereview.chromium.org/3099001

git-svn-id: http://src.chromium.org/svn/trunk/src@55644 4ff67af0-8c30-449e-8e8b-ad334ec8d88c