Stretch first run bubble linger time out to 3 seconds to avoid getting burned by page loading issues on slow Windows boxes.
This only affects the "Search from the omnibox" bubble that appears on the very first install of Chrome.

This solution came after a long discussion w/jcivelli yesterday.  A perfect solution to this problem would be too complex for M6, because it would involve triggering only when all first run pages have actually finished loading in the browser; this solution manages to both avoid getting stepped on by loading in all but the slowest machines -- and, if it does get stepped on, to at least actually show the bubble for 3 seconds.

BUG=51704
TEST=see bug
Review URL: http://codereview.chromium.org/3141015

git-svn-id: http://src.chromium.org/svn/trunk/src@56078 4ff67af0-8c30-449e-8e8b-ad334ec8d88c