Relanding r55510:
This CL originally caused crashes on the bots. I tracked the problem down to a bug in ANGLE that I fixed upstream:

http://codereview.appspot.com/1956042/show

This fix was committed upstream and I rolled the Chrome DEPS so the bot crash issue should be resolved.



Original comment for r55510:

Prelaunch GPU process after 10 seconds.

This will happen if the --prelaunch-gpu-process command line switch is present
or for 5% of users in a field trial.

This is so we can collect some stats of GPU capabilities.


TEST=try
BUG=none

Review URL: http://codereview.chromium.org/3179010

git-svn-id: http://src.chromium.org/svn/trunk/src@56239 4ff67af0-8c30-449e-8e8b-ad334ec8d88c