Fix windows doc generation to use depot_tools python.
The original bug report was that on windows, in the newly added directory.py 
file (for autogenerating samples) the 'import hashlib' line was failing because
that library isn't present in python 2.4.

To fix this, I've replaced the build.bat script with build.sh, which finds the 
depot_tools path and uses the python from there. This means you can just run
this from inside a cygwin bash shell instead of needing to run in a windows
cmd.exe.

Also fix an error message in build.py about missing os.EX_OK (it isn't 
guaranteed to be available on non-windows platforms).

BUG=52463
TEST=Running the new build.sh script should work on windows


Review URL: http://codereview.chromium.org/3157020

git-svn-id: http://src.chromium.org/svn/trunk/src@56437 4ff67af0-8c30-449e-8e8b-ad334ec8d88c