Start/stop service process when browser starts and stop
Save the information that the setup of remoting has completed.

After setup has been completed we start and stop service process when
--enable-remoting presents when browser starts.

Also save the information in the service process that host is registered
and ready to be used. Again start chromoting host automatically once
the host registration is done.

Review URL: http://codereview.chromium.org/3185015

git-svn-id: http://src.chromium.org/svn/trunk/src@56600 4ff67af0-8c30-449e-8e8b-ad334ec8d88c