Fix a few minor issues.
1 is that Client objects are added to the 
global list of clients before they are ready.

I just added a check that they are ready with
if (this.gl) although we could change the code
so they don't get added in the constructor
but instead in initWithCanvas.

Also added "webgl" as a context name.


TEST=none
BUG=none

Review URL: http://codereview.chromium.org/3164044

git-svn-id: http://src.chromium.org/svn/trunk/src@57613 4ff67af0-8c30-449e-8e8b-ad334ec8d88c