Fix data race in destroying the info bubble.
The info bubble is a member variable of the class and the controller class's lifetime is managed in the IO thread. However the info bubble's lifetime is managed in the UI thread and in this case it was not destroyed before the IO thread was notified of the event. So the IO thread could process the notification and destroy the controller object before it finished with releasing the bubble. This change destroys the bubble before notifying the delegate to prevent such cases.

BUG=53765
TEST=running unit_tests --gtest_filter="*Speech*Cancelled*" under RaceChecker should find no data races.

Review URL: http://codereview.chromium.org/3328001

git-svn-id: http://src.chromium.org/svn/trunk/src@58169 4ff67af0-8c30-449e-8e8b-ad334ec8d88c