Implement the speech input bubble on linux/gtk.
This is a popup which gets displayed below an <input> element when the user starts speech recognition on it. It displays a recognition status icon, an instruction label ('Speak now') at the top and a Cancel button at the bottom. The speech input module gets callbacks on user actions such as clicking Cancel or clicking any other window, and on both cases recognition stops and the bubble closes.

As part of this CL I had to move a constant from the windows code to the class so it can be used in gtk and mac.

TEST=manual, the code to invoke the bubble depends on another pending CL.
BUG=53598

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=58095

Review URL: http://codereview.chromium.org/3231009

git-svn-id: http://src.chromium.org/svn/trunk/src@58221 4ff67af0-8c30-449e-8e8b-ad334ec8d88c