Fix DeviceOrientationProviderTest.ObserverNotRemoved.
The problem was that some extra reference to Provider was left after the
ObserverNotRemoved test had finished. This reference was held by the
RunnableMethod object that executes ProviderImpl::DoInitializePollingThread on
a background thread. In the test, there were no guarantees that the body of
this function would finish before the test ended, thus causing a problem.

Fixing this by pushing another orientation through, forcing the Provider to
execute at least one Poll task on the background thread, which can only happen
after the initialization of the polling thread is done.

BUG=53865
TEST=unit_tests --gtest_filter=DeviceOrientationProviderTest.ObserverNotRemoved

Review URL: http://codereview.chromium.org/3232005

git-svn-id: http://src.chromium.org/svn/trunk/src@58359 4ff67af0-8c30-449e-8e8b-ad334ec8d88c