Revert "Make AutomationProvider fail fast on messages it doesn't understand."
Original review: http://codereview.chromium.org/3224001

The mentioned bug involves a crash on chromebot on Mac.  My working
hypothesis, based on some of the log messages seen, is that some file
descriptors are being trashed.  The CL being reverted closes the
channel while handling a message, which seems like it has potential to
be a problem.

Since this should not have user-facing impact, I am reverting the
change for merging to the 517 branch, after which I will re-instate
the change.  We have been trying to verify that this change is in the
crashing set, or out, but our data turned out to be incorrect, so I'm
doing this in parallel with a new run.

BUG=54453
TEST=chromebot

Review URL: http://codereview.chromium.org/3317023

git-svn-id: http://src.chromium.org/svn/trunk/src@59260 4ff67af0-8c30-449e-8e8b-ad334ec8d88c