Fix SessionStorage
Apparently the session storage code was pretty horribly broken.  It didn't correctly handle tabs being restored, didn't have the proper lifetime (this was the issue exposed in the bug), and had many leaks.

To fix this, things had to be plumbed fairly differently.  We need to pass session storage in on TabContents creation to ensure that the first RenderView will have the correct session storage id.  When closing a tab, we need to save the session storage with the restoration service.  When restoring a tab, we pass it back into the tab contents class.  When duplicating a tab, we clone the storage.

Lifetimes are now handled by standard reference counting code.  A SessionStorageNamespace object wraps an ID.  When it's instantiated, it allocates an ID.  When it's destroyed, it deletes the ID.  IDs make this process very lightweight (the heavyweight stuff is allocated on first use of SessionStorage) and it seperates the more complex lifetimes of SessionStorage namespaces (where less deterministic shutdown is more tollerable) from the LocalStorage namespace which needs to shutdown very precisely.

BUG=52393
TEST=Set some variable on session storage; close the tab; re-open the tab; the variable should still be set.  You can also run through the repro steps in the bug.

Review URL: http://codereview.chromium.org/3325012

git-svn-id: http://src.chromium.org/svn/trunk/src@59350 4ff67af0-8c30-449e-8e8b-ad334ec8d88c