Show "DNS" as the authority for DNSSEC validated certficiates.
At the moment, if we validate a certificate using DNSSEC the Page Info
dialog will show whatever Issuer the site chose as the issuing
authority. That's confusing because the site could choose a string
like 'US Dept of Homeland Security' (for example) which is misleading.

This patch forces the authority string to always be "DNSSEC" in the
case that we used DNSSEC to validate the certificate.

(The string "DNSSEC" isn't translated as it's an acronym.)

BUG=none
TEST=none

http://codereview.chromium.org/3304016/show

git-svn-id: http://src.chromium.org/svn/trunk/src@59686 4ff67af0-8c30-449e-8e8b-ad334ec8d88c