Support for building Chrome using Clang.
To build, set the clang=1 gyp_define.

This patch is the culmination of many months of effort and many
patches.  It contains the minimal changes to Chrome that
are Clang-specific.

With this, I can build the "chrome" target.  Once this patch
is in, we can incrementally fix bits of Chrome and various
tests and remove the Clang-specific workarounds.

Review URL: http://codereview.chromium.org/522020

git-svn-id: http://src.chromium.org/svn/trunk/src@59882 4ff67af0-8c30-449e-8e8b-ad334ec8d88c