NSS: Snap Start NPN support and other fixes.
Snap Start now works with NPN.

In the case that the server doesn't support Snap Start,
SSL_GetPredictedServerHelloData will no longer return any data to
allow the application not to cache Snap Start information for servers
that don't support it.

Previously Snap Start wasn't filling in the certificates in all cases.
(Test cases didn't check the certificates on the socket; but
Chrome noticed.)

BUG=none
TEST=nsstests (not in tree)

http://codereview.chromium.org/3430007/show

git-svn-id: http://src.chromium.org/svn/trunk/src@59957 4ff67af0-8c30-449e-8e8b-ad334ec8d88c