Add a PhishingClassifier implementation.
The PhishingClassifier ties together the feature extractors and the scorer,
and will be invoked by the RenderView when a page finishes loading.  This
change also makes the feature extractors have a non-owned pointer to the clock
so that it can be shared, and fixes a problem where the DOM feature extractor
cleared out any already-present features.

BUG=none
TEST=PhishingClassifierTest

Review URL: http://codereview.chromium.org/3452010

git-svn-id: http://src.chromium.org/svn/trunk/src@60082 4ff67af0-8c30-449e-8e8b-ad334ec8d88c