Eliminate HttpStreamHandle.  The name confused me.
HttpStreamHandle was a combination of an HttpStream and a
scoped_ptr<ClientSocketHandle>.  This let it manage the transport socket if
so desired.  I think that the HttpStream should be in charge of managing this.
* HttpBasicStream should always release it to the pool when done, but perhaps
  disconnect it first.
* HttpPipelinedStream (or whatever we name it) should know not to disconnect
  the transport socket or whatever.  It should return it to the pipeline stream
  manager.
* SPDY subtypes of HttpStream do not manage the transport socket.  They let the
  SpdySession handle it.
Since the ownership pattern varies based on the HttpStream subtype, I think
letting a HttpStreamHandle class perhaps control it is confusing.  It's better
for the subtype to know what it's supposed to do.

There was only one hangup here, the HttpProxyClientSocket, since it might need
to Disconnect() and then re-Connect() the transport socket.  It was using an
HttpBasicStream, which, with my change, would own the transport socket handle.
I fixed this by making the HttpProxyClientSocket create an HttpStreamParser
instead, which does not own the transport socket handle.

Review URL: http://codereview.chromium.org/3133029

git-svn-id: http://src.chromium.org/svn/trunk/src@60117 4ff67af0-8c30-449e-8e8b-ad334ec8d88c